infra/ansible/requirements.txt
ansible-core>=2.17,<2.19
pypsrp>=0.9.1,<1.0.0
pywinrm>=0.5.0
pyspnego[ntlm]>=0.10.0
requests-credssp>=2.0.0
pexpect>=4.9.0
pykerberos>=1.2.4
ldap3>=2.9.1
